refactor(conversion): aggregate conversion pass defs in Conversion/Passes.td
Follow the upstream MLIR layout (mlir/Conversion/Passes.td + Passes.h): put the
pass `def` in a single aggregate include/hip/Conversion/Passes.td instead of a
per-conversion Passes.td, and register via GEN_PASS_REGISTRATION in the
aggregate include/hip/Conversion/Passes.h. The per-conversion directory now only
holds a header that pulls its own declarations via GEN_PASS_DECL_<NAME>.

- Move def ConvertOnnxToHipsrPass to include/hip/Conversion/Passes.td; tablegen
  it into include/hip/Conversion/Passes.h.inc (target renamed
  HipsrConversionPassesIncGen -> HipConversionPassIncGen).
- Conversion/Passes.h: add GEN_PASS_REGISTRATION over the aggregate .inc.
- Conversion/OnnxToHipsr/Passes.h: keep only GEN_PASS_DECL for this pass (future
  populate*Patterns helpers go here too).
- Point OnnxToHipsr.cpp's GEN_PASS_DEF include and all IncGen deps at the
  aggregate target.

No functional change: convert-onnx-to-hipsr registers as before, LIT green.
