+++ include/hip/Conversion/CMakeLists.txt
@@ -3,4 +3,8 @@
 # ** Licensed under the MIT License.
 ##
 
-add_subdirectory(OnnxToHipsr)
+# Aggregate conversion-pass declarations/registration (upstream MLIR layout:
+# one Passes.td / Passes.h.inc for all TableGen-based conversions).
+set(LLVM_TARGET_DEFINITIONS Passes.td)
+mlir_tablegen(Passes.h.inc -gen-pass-decls -name OnnxToHipsr)
+add_public_tablegen_target(HipConversionPassIncGen)
