--- include/hip/Conversion/OnnxToHipsr/CMakeLists.txt
@@ -1,8 +1,0 @@
-##
-# ** Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
-# ** Licensed under the MIT License.
-##
-
-set(LLVM_TARGET_DEFINITIONS Passes.td)
-mlir_tablegen(Passes.h.inc -gen-pass-decls -name OnnxToHipsr)
-add_public_tablegen_target(HipsrConversionPassesIncGen)
