+++ lib/Conversion/OnnxToHipsr/CMakeLists.txt
@@ -8,7 +8,7 @@ add_library(OnnxToHipsr STATIC
 )
 
 add_dependencies(OnnxToHipsr
-  HipsrConversionPassesIncGen
+  HipConversionPassIncGen
   HipsrDialectIncGen
 )
