+++ tools/hip-mlir-opt/CMakeLists.txt
@@ -8,7 +8,7 @@ add_executable(hip-mlir-opt hip-mlir-opt.cpp)
 add_dependencies(hip-mlir-opt
   HipDialectIncGen
   HipsrDialectIncGen
-  HipsrConversionPassesIncGen
+  HipConversionPassIncGen
   HipTypesIncGen
   HipOpsIncGen
   HipPassesIncGen
